requirements.txt
# Core ML
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.49.0
accelerate>=0.34.0
qwen-vl-utils>=0.0.10

# Video processing
opencv-python-headless>=4.10.0
Pillow>=10.4.0
av>=12.0.0
decord>=0.6.0

# Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# Utils
python-dotenv>=1.0.1
pydantic>=2.8.0
numpy>=1.26.0,<2.0.0
psutil>=6.0.0